move ResourceRequirement to a separate file
