+++ operator/v1/funcs_clustermanager.go
@@ -1,0 +1,5 @@
+package v1
+
+func (cm *ClusterManager) GetResourceRequirement() *ResourceRequirement {
+	return cm.Spec.ResourceRequirement
+}
