+++ operator/v1/funcs_klusterlet.go
@@ -1,0 +1,5 @@
+package v1
+
+func (k *Klusterlet) GetResourceRequirement() *ResourceRequirement {
+	return k.Spec.ResourceRequirement
+}
